+++ private/language/english_utf-8.php
@@ -257,6 +257,7 @@ $LANG01 = array(
     'blocks_left_title'         => 'Left Side',
     'blocks_right_title'        => 'Right Side',
     'blocks_footer_title'       => 'Footer',
+    'source'                    => 'Source',
 );
 
 ###############################################################################
@@ -1009,6 +1010,10 @@ $LANG24 = array(
     102 => 'Attribute',
     103 => 'Value',
     104 => 'Update',
+    105 => 'Attribution Source URL',
+    106 => 'Attribution Source Site Name',
+    107 => 'Attribution Author',
+    108 => 'Attribution',
 );
 
 ###############################################################################