+++ public_html/admin/install/include/install.lib.php
@@ -1345,9 +1345,11 @@ function INST_doDatabaseUpgrades($current_fusion_version, $use_innodb = false)
             $c->del('mysqldump_options','Core');
 
             $_SQL[] = "ALTER TABLE {$_TABLES['blocks']} CHANGE `title` `title` VARCHAR(255) NULL DEFAULT NULL;";
+            $_SQL[] = "ALTER TABLE {$_TABLES['stories']} ADD `attribution_url` VARCHAR(255) NOT NULL default '' AFTER `expire`;";
+            $_SQL[] = "ALTER TABLE {$_TABLES['stories']} ADD `attribution_name` VARCHAR(255) NOT NULL DEFAULT '' AFTER `attribution_url`;";
+            $_SQL[] = "ALTER TABLE {$_TABLES['stories']} ADD `attribution_author` VARCHAR(255) NOT NULL DEFAULT '' AFTER `attribution_name`;";
 
             list($rc,$errors) = INST_updateDB($_SQL);
-
 
             $current_fusion_version = '1.6.1';